requirements.txt
yaml
numpy
grpcio>=1.12.0
tensorflow>=1.4.0
grpcio-tools>=1.12.0
